api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic==0.40.0
python-dotenv==1.0.0
requests==2.31.0
recipe-scrapers==14.51.0
Pillow==10.0.0
beautifulsoup4==4.12.0
certifi==2023.7.22
stripe>=8.0.0
werkzeug>=3.0.0
